fix doc/warning messages and add valve module to documentation
